amplify/agent/runtime/requirements.txt
bedrock-agentcore
strands-agents
tavily-python
